rag-engine/requirements.txt
# Python 3.12+ 호환

# 핵심 AI 프레임워크
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# 벡터 데이터베이스
chromadb>=0.4.0

# API 서버
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 문서 처리
pypdf2>=3.0.1
python-docx>=0.8.11
beautifulsoup4>=4.12.2
lxml>=4.9.3

# HTTP 클라이언트
requests>=2.31.0
aiohttp>=3.9.0

# 데이터 모델 및 설정
pydantic>=2.5.0
python-dotenv>=1.0.0

# 유틸리티
typing-extensions>=4.8.0
python-json-logger>=2.0.7

# 개발/테스트 (선택적)
pytest>=7.4.0
pytest-asyncio>=0.21.0